requirements.txt
requests>=2.28.0
pydantic>=2.0.0
click>=8.0.0
rich>=13.0.0
pyyaml>=6.0
